feat(topic): add topic settings editing ui
