+++ shared/src/commonMain/kotlin/org/nsh07/pomodoro/ui/settingsScreen/screens/TimerSettings.kt
@@ -24,7 +24,6 @@ import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -40,8 +39,6 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.foundation.text.input.rememberTextFieldState
 import androidx.compose.material3.ButtonDefaults
@@ -60,10 +57,10 @@ import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialShapes
 import androidx.compose.material3.MaterialTheme.colorScheme
 import androidx.compose.material3.MaterialTheme.motionScheme
+import androidx.compose.material3.MaterialTheme.shapes
 import androidx.compose.material3.MaterialTheme.typography
 import androidx.compose.material3.MenuDefaults
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Slider
 import androidx.compose.material3.SliderState
 import androidx.compose.material3.SplitButtonDefaults
 import androidx.compose.material3.SplitButtonLayout
@@ -85,7 +82,6 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -97,12 +93,10 @@ import org.jetbrains.compose.resources.stringResource
 import org.nsh07.pomodoro.data.Topic
 import org.nsh07.pomodoro.data.Topic.Companion.defaultTopic
 import org.nsh07.pomodoro.ui.mergePaddingValues
-import org.nsh07.pomodoro.ui.rememberRequestDndPermissionCallback
 import org.nsh07.pomodoro.ui.settingsScreen.SettingsSwitchItem
-import org.nsh07.pomodoro.ui.settingsScreen.components.MinuteInputField
-import org.nsh07.pomodoro.ui.settingsScreen.components.MinutesInputTransformation3Digits
 import org.nsh07.pomodoro.ui.settingsScreen.components.PlusDivider
 import org.nsh07.pomodoro.ui.settingsScreen.components.SliderListItem
+import org.nsh07.pomodoro.ui.settingsScreen.components.TopicTimerSettings
 import org.nsh07.pomodoro.ui.settingsScreen.viewModel.SettingsAction
 import org.nsh07.pomodoro.ui.settingsScreen.viewModel.SettingsState
 import org.nsh07.pomodoro.ui.theme.CustomColors.detailPaneTopBarColors
@@ -125,33 +119,22 @@ import tomato.shared.generated.resources.always_on_display_desc
 import tomato.shared.generated.resources.aod
 import tomato.shared.generated.resources.arrow_back
 import tomato.shared.generated.resources.arrow_down
-import tomato.shared.generated.resources.auto_start_next_timer
-import tomato.shared.generated.resources.auto_start_next_timer_desc
-import tomato.shared.generated.resources.autoplay
 import tomato.shared.generated.resources.back
 import tomato.shared.generated.resources.check
 import tomato.shared.generated.resources.clear
-import tomato.shared.generated.resources.clocks
 import tomato.shared.generated.resources.daily_focus_goal
-import tomato.shared.generated.resources.dnd
-import tomato.shared.generated.resources.dnd_desc
 import tomato.shared.generated.resources.flag
-import tomato.shared.generated.resources.focus
 import tomato.shared.generated.resources.hours_and_minutes_format
 import tomato.shared.generated.resources.info
-import tomato.shared.generated.resources.long_break
 import tomato.shared.generated.resources.mobile_lock_portrait
 import tomato.shared.generated.resources.new_label
 import tomato.shared.generated.resources.pomodoro_info
 import tomato.shared.generated.resources.secure_aod
 import tomato.shared.generated.resources.secure_aod_desc
-import tomato.shared.generated.resources.session_length
-import tomato.shared.generated.resources.session_length_desc
 import tomato.shared.generated.resources.session_only_progress
 import tomato.shared.generated.resources.session_only_progress_desc
 import tomato.shared.generated.resources.settings
 import tomato.shared.generated.resources.settings_infinite_focus_tip
-import tomato.shared.generated.resources.short_break
 import tomato.shared.generated.resources.timer
 import tomato.shared.generated.resources.timer_settings_reset_info
 import tomato.shared.generated.resources.view_day
@@ -181,54 +164,28 @@ fun TimerSettings(
         .windowSizeClass
         .isWidthAtLeastBreakpoint(WIDTH_DP_EXPANDED_LOWER_BOUND)
 
-    val requestDndPermissionCallback = rememberRequestDndPermissionCallback()
-
     val switchItems = remember(
-        editingTopic.dndEnabled,
         settingsState.aodEnabled,
-        editingTopic.autostartNextSession,
         settingsState.secureAod,
         isPlus,
         serviceRunning
     ) {
         listOf(
-            listOf(
-                SettingsSwitchItem(
-                    checked = editingTopic.autostartNextSession,
-                    icon = Res.drawable.autoplay,
-                    label = Res.string.auto_start_next_timer,
-                    description = Res.string.auto_start_next_timer_desc,
-                    onClick = { onAction(SettingsAction.SaveAutostartNextSession(it)) }
-                ),
-                SettingsSwitchItem(
-                    checked = editingTopic.dndEnabled,
-                    enabled = !serviceRunning,
-                    icon = Res.drawable.dnd,
-                    label = Res.string.dnd,
-                    description = Res.string.dnd_desc,
-                    onClick = {
-                        requestDndPermissionCallback(it)
-                        onAction(SettingsAction.SaveDndEnabled(it))
-                    }
-                )
+            SettingsSwitchItem(
+                checked = settingsState.aodEnabled,
+                enabled = isPlus,
+                icon = Res.drawable.aod,
+                label = Res.string.always_on_display,
+                description = Res.string.always_on_display_desc,
+                onClick = { onAction(SettingsAction.SaveAodEnabled(it)) }
             ),
-            listOf(
-                SettingsSwitchItem(
-                    checked = settingsState.aodEnabled,
-                    enabled = isPlus,
-                    icon = Res.drawable.aod,
-                    label = Res.string.always_on_display,
-                    description = Res.string.always_on_display_desc,
-                    onClick = { onAction(SettingsAction.SaveAodEnabled(it)) }
-                ),
-                SettingsSwitchItem(
-                    checked = settingsState.secureAod && isPlus,
-                    enabled = isPlus && settingsState.aodEnabled,
-                    icon = Res.drawable.mobile_lock_portrait,
-                    label = Res.string.secure_aod,
-                    description = Res.string.secure_aod_desc,
-                    onClick = { onAction(SettingsAction.SaveSecureAod(it)) }
-                )
+            SettingsSwitchItem(
+                checked = settingsState.secureAod && isPlus,
+                enabled = isPlus && settingsState.aodEnabled,
+                icon = Res.drawable.mobile_lock_portrait,
+                label = Res.string.secure_aod,
+                description = Res.string.secure_aod_desc,
+                onClick = { onAction(SettingsAction.SaveSecureAod(it)) }
             )
         )
     }
@@ -414,103 +371,18 @@ fun TimerSettings(
                     Spacer(Modifier.height(14.dp))
                 }
                 item {
-                    Row(
-                        horizontalArrangement = Arrangement.Center,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .horizontalScroll(rememberScrollState())
-                    ) {
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                            verticalArrangement = Arrangement.spacedBy(2.dp)
-                        ) {
-                            Text(
-                                stringResource(Res.string.focus),
-                                style = typography.titleSmallEmphasized
-                            )
-                            MinuteInputField(
-                                state = focusTimeInputFieldState,
-                                enabled = !serviceRunning,
-                                shape = RoundedCornerShape(
-                                    topStart = topListItemShape.topStart,
-                                    bottomStart = topListItemShape.topStart,
-                                    topEnd = topListItemShape.bottomStart,
-                                    bottomEnd = topListItemShape.bottomStart
-                                ),
-                                inputTransformation = MinutesInputTransformation3Digits,
-                                imeAction = ImeAction.Next
-                            )
-                        }
-                        Spacer(Modifier.width(2.dp))
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                            verticalArrangement = Arrangement.spacedBy(2.dp)
-                        ) {
-                            Text(
-                                stringResource(Res.string.short_break),
-                                style = typography.titleSmallEmphasized
-                            )
-                            MinuteInputField(
-                                state = shortBreakTimeInputFieldState,
-                                enabled = !serviceRunning,
-                                shape = RoundedCornerShape(middleListItemShape.topStart),
-                                imeAction = ImeAction.Next
-                            )
-                        }
-                        Spacer(Modifier.width(2.dp))
-                        Column(
-                            horizontalAlignment = Alignment.CenterHorizontally,
-                            verticalArrangement = Arrangement.spacedBy(2.dp)
-                        ) {
-                            Text(
-                                stringResource(Res.string.long_break),
-                                style = typography.titleSmallEmphasized
-                            )
-                            MinuteInputField(
-                                state = longBreakTimeInputFieldState,
-                                enabled = !serviceRunning,
-                                shape = RoundedCornerShape(
-                                    topStart = bottomListItemShape.topStart,
-                                    bottomStart = bottomListItemShape.topStart,
-                                    topEnd = bottomListItemShape.bottomStart,
-                                    bottomEnd = bottomListItemShape.bottomStart
-                                ),
-                                imeAction = ImeAction.Done
-                            )
-                        }
-                    }
+                    TopicTimerSettings(
+                        topic = editingTopic,
+                        serviceRunning = serviceRunning,
+                        focusTimeInputFieldState = focusTimeInputFieldState,
+                        shortBreakTimeInputFieldState = shortBreakTimeInputFieldState,
+                        longBreakTimeInputFieldState = longBreakTimeInputFieldState,
+                        sessionsSliderState = sessionsSliderState,
+                        onAction = onAction
+                    )
+                    Spacer(Modifier.height(14.dp))
                 }
-                item {
-                    Spacer(Modifier.height(12.dp))
-                }
-                item {
-                    Column(Modifier.background(listItemColors.containerColor, topListItemShape)) {
-                        ListItem(
-                            leadingContent = {
-                                Icon(painterResource(Res.drawable.clocks), null)
-                            },
-                            headlineContent = {
-                                Text(stringResource(Res.string.session_length))
-                            },
-                            supportingContent = {
-                                Text(
-                                    stringResource(
-                                        Res.string.session_length_desc,
-                                        sessionsSliderState.value.toInt()
-                                    )
-                                )
-                            },
-                            colors = listItemColors,
-                            modifier = Modifier.clip(cardShape)
-                        )
-                        Slider(
-                            state = sessionsSliderState,
-                            enabled = !serviceRunning,
-                            modifier = Modifier
-                                .padding(start = (16 * 2 + 24).dp, end = 16.dp, bottom = 12.dp)
-                        )
-                    }
-                }
+
                 item {
                     val hmf = stringResource(Res.string.hours_and_minutes_format)
                     SliderListItem(
@@ -520,7 +392,7 @@ fun TimerSettings(
                         label = stringResource(Res.string.daily_focus_goal),
                         trailingLabel = { millisecondsToHoursMinutes(it.toLong(), hmf) },
                         icon = { Icon(painterResource(Res.drawable.flag), null) },
-                        shape = bottomListItemShape
+                        shape = shapes.large
                     ) {
                         with(it.toLong()) {
                             onAction(SettingsAction.SaveFocusGoal(this - (this % (30 * 60 * 1000))))
@@ -529,54 +401,9 @@ fun TimerSettings(
                 }
                 item { Spacer(Modifier.height(12.dp)) }
 
-                itemsIndexed(switchItems[0]) { index, item ->
-                    ListItem(
-                        leadingContent = {
-                            Icon(
-                                painterResource(item.icon),
-                                contentDescription = null,
-                                modifier = Modifier.padding(top = 4.dp)
-                            )
-                        },
-                        headlineContent = { Text(stringResource(item.label)) },
-                        supportingContent = { Text(stringResource(item.description)) },
-                        trailingContent = {
-                            Switch(
-                                checked = item.checked,
-                                enabled = item.enabled,
-                                onCheckedChange = { item.onClick(it) },
-                                thumbContent = {
-                                    if (item.checked) {
-                                        Icon(
-                                            painter = painterResource(Res.drawable.check),
-                                            contentDescription = null,
-                                            modifier = Modifier.size(SwitchDefaults.IconSize),
-                                        )
-                                    } else {
-                                        Icon(
-                                            painter = painterResource(Res.drawable.clear),
-                                            contentDescription = null,
-                                            modifier = Modifier.size(SwitchDefaults.IconSize),
-                                        )
-                                    }
-                                },
-                                colors = switchColors
-                            )
-                        },
-                        colors = listItemColors,
-                        modifier = Modifier.clip(
-                            when (index) {
-                                0 -> topListItemShape
-                                switchItems[0].size - 1 -> bottomListItemShape
-                                else -> middleListItemShape
-                            }
-                        )
-                    )
-                }
-
                 if (isPlus) {
                     item { Spacer(Modifier.height(12.dp)) }
-                    itemsIndexed(switchItems[1]) { index, item ->
+                    itemsIndexed(switchItems) { index, item ->
                         ListItem(
                             leadingContent = {
                                 Icon(
@@ -614,7 +441,7 @@ fun TimerSettings(
                             modifier = Modifier.clip(
                                 when (index) {
                                     0 -> topListItemShape
-                                    switchItems[1].size - 1 -> bottomListItemShape
+                                    switchItems.size - 1 -> bottomListItemShape
                                     else -> middleListItemShape
                                 }
                             )
@@ -680,7 +507,7 @@ fun TimerSettings(
                     item {
                         PlusDivider(setShowPaywall)
                     }
-                    itemsIndexed(switchItems[1]) { index, item ->
+                    itemsIndexed(switchItems) { index, item ->
                         ListItem(
                             leadingContent = {
                                 Icon(
@@ -718,7 +545,7 @@ fun TimerSettings(
                             modifier = Modifier.clip(
                                 when (index) {
                                     0 -> topListItemShape
-                                    switchItems[1].size - 1 -> bottomListItemShape
+                                    switchItems.size - 1 -> bottomListItemShape
                                     else -> middleListItemShape
                                 }
                             )
